Site Receita

+++ php/SistemaReceitas/pages/deletar.php
@@ -1,6 +1,11 @@
 <?php
 include '../database/db.php';
-$id = $_GET['id'];
-$conn->query("DELETE FROM receitas WHERE id=$id");
-header('Location: ../index.php');
-?>
+
+if (isset($_GET['id'])) {
+    $id = intval($_GET['id']);
+    $conn->query("DELETE FROM receitas WHERE id = $id");
+}
+
+header("Location: ../index.php");
+exit;
+
